report a wrong city name
when the city is not found, change the error to a latitude/longitude value equal to 0. then  a test can be done to see if the city does exist or not. without this and with the simple error report, a red message is reported in the console but can't de dicted to the TTS to infor the end user